requirements.txt
toml
groq
python-docx
pydantic-ai
fastapi[all]
uvicorn
